+++ examples/data-sources/catalystcenter_update_authentication_profile/data-source.tf
@@ -1,4 +1,3 @@
 data "catalystcenter_update_authentication_profile" "example" {
-  id                          = "cc0b30bc-94e7-458a-80e2-c7bbecc829f5"
   authentication_profile_name = "Closed Authentication"
 }