+++ examples/resources/catalystcenter_assign_templates_to_tag/import.sh
@@ -1,0 +1,1 @@
+terraform import catalystcenter_assign_templates_to_tag.example "<tag_id>"